feat: Switch to wkd-exporter for exporting WKD directory
As sequoia-sq >= 0.38 breaks our use-case of exporting a keyring to a
WKD directory structure, we switch to the simpler `wkd-exporter`.

In addition to all `archlinux.org` User IDs, we now also export all
`master-key.archlinux.org` User IDs.

The `wkd-exporter` tool omits the creation of the `.well-known`
directory, as the dedicated `openpgp` directory in it is the relevant
bit. Therefore we adjust (the use of) `WKD_BUILD_DIR` accordingly.

Signed-off-by: David Runge <dvzrv@archlinux.org>