fix(tests): sequoia-sq 1.0 compat: avoid reusing file handles
Previously we opened a temporary file handle and have used this handle
in multiple stages during creating a revocation certification, which has
worked so far. Since latest sequoia, the passed file does not seem to be
truncated but unlinked before being written to, which means our cached
file handle is not valid anymore and subsequent truncation and write
operations will point into the void instead of the inode currently used
for the specific path. This lead to the actual certificate containing
the revocation to never be written to the expected path, which means the
keyring tooling has never picked up this packet.

Stop relying on this behavior and instead just write to a handle once
and create a new file for any further stages where we need to write
a certificate to the filesystem.

Signed-off-by: Levente Polyak <anthraxx@archlinux.org>