chore: Remove certificate of sudoforge
The certificate has been added for a new package maintainer.
Unfortunately the process was never finished and the certificate was
never revoked either.
As the individual ghosted us for years and the certificate also carries
no third-party certificates, we can safely remove it to not expose it on
new user systems any longer.

Fixes: https://gitlab.archlinux.org/archlinux/archlinux-keyring/-/issues/216
Signed-off-by: David Runge <dvzrv@archlinux.org>